+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,14 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import image_clipboard
+import pasteboard
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  ImageClipboardPlugin.register(with: registry.registrar(forPlugin: "ImageClipboardPlugin"))
+  PasteboardPlugin.register(with: registry.registrar(forPlugin: "PasteboardPlugin"))
+}