+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- Added support to read image from clipboard on Desktop.
+- Added support to add a web paste listener on Web.
+
 ## 1.0.0
 
 - Initial release.